Hybrid authentication definition
This may be useful if both components are entered at the same time, but in many implementations it is a sequential authentication process, and if the first component entered is incorrect, the system does not allow the user to try the second entry. So for example, if the first component is the biometric and the user is unable to pass the biometric, they will not be prompted to enter the password.

It seems the real key here is that a failure of either is a failure to authenticate, not that both must be entered and the reason for the failure not reported.