Change related to objectives becoming mandatory
The current sentence states that objectives will become mandatory in the "next" revision. In practice, it has happened at some point in the future, but not necessarily the "next" revision. This just points out that they will become mandatory at some point without being quite so specific about the when.