Remove GPS for just location
Many devices today are using more than just the GPS for location (Wi-Fi, Cellular, even BT). 

I think it should instead just say to retrieve the current location since that would cover the broader set of services that make up the location services on the device and not just whether or not GPS is enabled.